% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ASGS.R
\name{install_ASGS}
\alias{install_ASGS}
\title{Install a (nearly) complete package of the Australian Statistical Geography Standard}
\usage{
install_ASGS(temp.tar.gz = tempfile(fileext = ".tar.gz"), ...)
}
\arguments{
\item{temp.tar.gz}{A file to save the ASGS tarball. Since the package is quite large,
it may be prudent to set this to a non-temporary file so that you can attempt reinstallation.}

\item{...}{Arguments passed to \code{\link[utils]{install.packages}}.}
}
\description{
Install a (nearly) complete package of the Australian Statistical Geography Standard
}
