% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-STE_2016.R
\docType{data}
\name{STE_2016_simple}
\alias{STE_2016_simple}
\title{State shapefile}
\format{An object of class \code{SpatialPolygonsDataFrame} with 9 rows and 4 columns.}
\usage{
STE_2016_simple
}
\description{
State shapefile
}
\keyword{datasets}
